extract filter wrapping structs to separate file
